pipeline-registry/orchestrator/requirements.txt
anthropic>=0.40.0
playwright>=1.45.0
Pillow>=10.0.0
requests>=2.31.0
rich>=13.0.0
python-dotenv>=1.0.0
